+++ PlanBuild/PlanBuildPlugin.cs
@@ -75,7 +75,8 @@ namespace PlanBuild
             }
 
             // BP Market GUI is OK in the main menu
-            if (BlueprintGUI.IsAvailable() && !SelectionSaveGUI.IsVisible() && !TerrainModGUI.IsVisible() &&
+            if (BlueprintGUI.IsAvailable() && 
+                !SelectionSaveGUI.IsVisible() && !TerrainModGUI.IsVisible() && !SelectionGUI.IsVisible() &&
                 (PlanBuild.Config.AllowMarketHotkey.Value || SynchronizationManager.Instance.PlayerIsAdmin) &&
                 ZInput.GetButtonDown(PlanBuild.Config.MarketHotkeyButton.Name))
             {